Notebook draft to train the dogs-vs-cats model
This adds a draft notebook to train a model that classifies images of dogs and cats.

The original notebook can be found in: https://github.com/pytorch/serve/blob/master/examples/Workflows/dog_breed_classification/cat_dog_classification.ipynb. The nodebook added in this commit is almost the same as the original notebook. Some adaptations are planned for later so that the model is exported in either onnx format or openvino_ir format.