Finish draft of train-model notebook
The draft of the notebook for training the model is "finished". This means that the notebook can now be used to train the model for classifying images of cats and dogs, and the model can be exported.

The exported formats that were added to the notebook are as state_dict, ONNX format, and openvino IR format.